% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianDFM.R
\docType{package}
\name{BayesianDFM}
\alias{BayesianDFM}
\title{A package for running a Bayesian DFM}
\description{
The BayesianDFM package provides relevant functions for running a Bayesian
Dynamic Factor Model.
}
\section{BayesianDFM functions}{
 \itemize{ \item BVAR_Jeff \item compFstate
\item drawLam \item drawSig \item get_IC \item get_nowcast \item
multimoveGibbs \item prepare_data \item run_model }
}

