% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BVAR_Jeff.R
\name{BVAR_Jeff}
\alias{BVAR_Jeff}
\title{Draws posterior of VAR(p) Model with non-informative prior}
\usage{
BVAR_Jeff(Yts, p, const)
}
\arguments{
\item{Yts}{n x T matrix of data.}

\item{p}{the lag lenght of the VAR.}

\item{const}{A scalar, where const = 1 for model with intercept, const = 0 for model without intercept.}
}
\value{
A List with:
\itemize{
\item betar_all n x np+1 x R-burn matrices of coefficients.
\item Sigr_all n x n x R-burn matrices of variance covariance matrices.
}
}
\description{
\code{create_inventory} constructs companion matrix for VAR(p) model.
}
\details{
This function draws from the posterior of a VAR(p) model with non-informatibve prior
}
\examples{
yt <- as.matrix(t(Xmat))
k <- 2
n <- dim(yt)[1]
q <- 1
m <- k*q
Q <- as.matrix(diag(0.1,k))
R <- as.matrix(diag(n)*0.01)
alpha_0 <- matrix(0,m,1)
P_0 <- diag(m)
lambdasim <- matrix(rep(rnorm(n,0,1)*0.1,k), nrow = n, ncol = k, byrow = TRUE)
diag(lambdasim) <- 1
lambdasim[upper.tri(lambdasim)] <- 0
lambda <- lambdasim
Tt <- dim(yt)[2]
phi <- diag(rnorm(k,0,1))
const <- 0
ft <-  multimoveGibbs(yt,phi,Q,lambda,const,Tt,q,alpha_0,P_0,R)
param <- BVAR_Jeff(ft,q,0)
}
