% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_inventory.R
\name{create_inventory}
\alias{create_inventory}
\title{Create an inventory for time series}
\usage{
create_inventory(flows, stocks)
}
\arguments{
\item{flows}{A list of time series variable (flow variables).}

\item{stocks}{A list of time series variable (stock variables).}
}
\value{
A table
}
\description{
\code{create_inventory} constructs an inventory. It calculates the mean,
standard deviation of the various series and outputs its frequency, type, and name in a table.
}
\details{
The function reads a list of variables that are of class time series. For every list the function looks for the following things:
\itemize{
\item name()
\item frequency()
\item mean()
\item sd()
}
}
\examples{
inventory <- create_inventory(flows = data$flows, stocks = data$stocks)
}
