% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_final.R
\docType{data}
\name{dat_final}
\alias{dat_final}
\title{List of indicators.}
\format{
A list containing of time series objects, demeaned and standardized:
}
\source{
datastream
}
\usage{
dat_final
}
\description{
A dataset containing of demeaned and normalized relevant indicators for the UK.
}
\keyword{datasets}
