% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{Stock and Flow list of indicators.}
\format{
A list containing of two list, each containing of time series objects:
\describe{
\item{stock}{indicators that are stock variables}
\item{flow}{indicators that are flow variables}
...
}
}
\source{
datastream
}
\usage{
data
}
\description{
A dataset containing of two list (stock, flow) containing relevant indicators for the UK.
}
\keyword{datasets}
