% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_Lam.R
\name{draw_Lam}
\alias{draw_Lam}
\title{Draw factor loadings from their conditional posterior}
\usage{
draw_Lam(yt, ft, R, lam0, V_lam)
}
\arguments{
\item{yt}{A matrix of demeaned and standardized time series data.}

\item{ft}{A matrix of the factors.}

\item{R}{Diagonal matrix of dimension n of idiosyncratic component.}

\item{lam0}{Matrix of mean of initial factor loadings.}

\item{V_lam}{Diagonal matrix of dimension k of Variance-Covariance matrix of
initial factor loadings.}
}
\value{
Element of a vector, coefficient of lambda.
}
\description{
function that allows to draw from posterior of factor loadings.
}
\details{
The function conducts a normalization by demeaning and dividing by the
standard deviation of the individual time series.
}
\examples{
k <- 2
lam0 <- matrix(0,k,1)
V_lam <- diag(k)
}
