% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawSig.R
\name{drawSig}
\alias{drawSig}
\title{Draw variance of idiosyncratic components from inverse gamma distribution}
\usage{
drawSig(yt, lambda, ft, Ttq, nu0, s0)
}
\arguments{
\item{yt}{A matrix of demeaned and standardized time series data.}

\item{lambda}{A vector of dimension n x k of the factor loadings.}

\item{ft}{A matrix of the factors.}

\item{Ttq}{Number of high-frequency periods minus lag length for state
equation.}

\item{nu0}{shape (degrees of freedom) of IG distribution.}

\item{s0}{Prior for the rate (reciprocal of scale) of IG distribution
(1/100=0,01 as benchmark residuals of AR(1) from OLS).}
}
\value{
Matrix R.
}
\description{
function  draws variance of idiosyncratic components from inverse gamma
distribution.
}
\examples{
yt <- as.matrix(t(Xmat))
k <- 2
q <- 1
Tt <- dim(yt)[2]
Ttq <- Tt-q
V_lam <- diag(k)

nu0 <- k + 3
s0 <- 100
}
