% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nowcast.R
\name{get_nowcast}
\alias{get_nowcast}
\title{Get nowcast from the bayesian DFM}
\usage{
get_nowcast(
  Xmat,
  s,
  q,
  alpha_0,
  P_0,
  inventory,
  target,
  flows,
  lambda_mean,
  phi_mean,
  R_mean,
  Q_mean,
  const
)
}
\arguments{
\item{Xmat}{Demeaned and standardized matrix of time series data.}

\item{s}{Number of periods for aggregation rule.}

\item{q}{Lag length for state equation.}

\item{alpha_0}{Vector of dimension m x 1 (Initial conditions for Kalman filter).}

\item{P_0}{Diagonal matrix of dimension m (Initial conditions for Kalman filter).}

\item{inventory}{Inventory of the corresponding data.}

\item{target}{Defines the target variable.}

\item{flows}{A list of time series variable (flow variables).}

\item{lambda_mean}{A vector of dimension n x k of the posterior mean factor loadings.}

\item{phi_mean}{Diagonal matrix of dimension k x k with vector of posterior mean autoregressive coefficients.}

\item{R_mean}{Diagonal matrix of dimension n of posterior mean idiosyncratic component.}

\item{Q_mean}{matrices of posterior mean variance covariance matrices.}

\item{const}{A scalar, where const = 1 for model with intercept, const = 0 for model without intercept.}
}
\value{
A time series of the target variable including forecast.
}
\description{
function  extends the data matrix by the forecasting period, draw extended factor conditional on posterior parameters and makes forecast.
}
\examples{
Xmat <- prepare_data(flows = data$flows, stocks = data$stocks, inventory, target = target)
s <- 2*(k - 1)
q <- 1
alpha_0 <- matrix(0,m,1)
P_0 <- diag(m)
inventory <- create_inventory(flows = data$flows, stocks = data$stocks)
target <- c("UKGDPM.YQ")
flows = data$flows
out$lambda_mean <- apply(simplify2array(out$lambda_all), 1:2, mean)
out$phi_mean <- apply(simplify2array(out$phi_all), 1:2, mean)
out$Q_mean <- apply(simplify2array(out$Q_all), 1:2, mean)
out$R_mean <- apply(simplify2array(out$R_all), 1:2, mean)
out$ncst <- get_nowcast(Xmat = Xmat, s = s, q = q, alpha_0 = alpha_0,
P_0 = P_0, inventory = inventory, target = target, flows = flows,
lambda_mean = out$lambda_mean, phi_mean = out$phi_mean, R_mean = out$R_mean,
Q_mean = out$Q_mean, const = const)
}
