% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{metadata}
\alias{metadata}
\title{Metadata of indicators.}
\format{
A data.frame with 23 rows and 12 variables:
\describe{
\item{keys}{tbd}
\item{name}{tbd}
\item{description}{tbd}
\item{source}{tbd}
\item{unit}{tbd}
\item{start.date}{tbd}
\item{category}{tbd}
\item{seasonal_adjustment}{tbd}
\item{release.date}{tbd}
\item{flow}{tbd}
\item{transform}{tbd}
\item{frequency}{tbd} }
}
\source{
datastream
}
\usage{
metadata
}
\description{
A table containing relevant metadata of the UK indicators.
}
\keyword{datasets}
