% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_model}
\alias{run_model}
\title{Get nowcast from the bayesian DFM}
\usage{
run_model(yt, k, q, m, n, Tt, Ttq, const, target, inventory, Xmat, flows)
}
\arguments{
\item{yt}{A matrix of demeaned and standardized time series data.}

\item{k}{Number of states (number of factors)}

\item{q}{Lag length for state equation.}

\item{m}{number of states (number of factors) x lag length for state equation (nxq)}

\item{n}{Number of variables}

\item{Tt}{Number of high-frequency periods.}

\item{Ttq}{Number of high-frequency periods minus lag length for state equation}

\item{const}{A scalar, where const = 1 for model with intercept, const = 0 for model without intercept.}

\item{target}{Defines the target variable.}

\item{inventory}{Inventory of the corresponding data.}

\item{Xmat}{A data set containing all the time series.}

\item{flows}{A list of time series variable (flow variables).}
}
\value{
A list with the mean and quantiles of the parameters, the forecast and the number of burns, draws and saves.
}
\description{
function extends the data matrix by the forecasting period, draws extended factor conditional on posterior parameters and makes forecast.
}
\examples{

out <- run_model(yt,k,q,m,n,Tt,Ttq,const,target)
}
