% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xmat.R
\docType{data}
\name{Xmat}
\alias{Xmat}
\title{Matrix of relvant indicators in time series format.}
\format{
A matrix with 367 rows and 23 variables:
\describe{
\item{rows}{time period}
\item{columns}{indicator}
...
}
}
\source{
datastream
}
\usage{
Xmat
}
\description{
A dataset containing relevant indicators in time series format for further analysis.
}
\keyword{datasets}
